Changed async to Enumerator for Web Build

+++ Assets/Scripts/Player/PlayerScriptableObject.cs
@@ -9,6 +9,6 @@ namespace CosmicCuration.Player
         public float movementSpeed;
         public float defaultFireRate;
         public float rapidFireRate;
-        public int deathDelay;
+        public float deathDelay;
     } 
 }
